backend/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
pydantic-settings
alembic
python-jose[cryptography]
passlib
bcrypt==3.2.2
python-dotenv
pdfplumber
httpx
python-multipart